refactor(client): move analytics slice into features/analytics
