+++ client/src/features/analytics/components/DashboardCharts.tsx
@@ -8,19 +8,19 @@ import {
   Layers,
   Building2,
 } from 'lucide-react';
-import { Card, CardContent, CardHeader, CardTitle } from '../../shared/ui/card';
-import { Button } from '../../shared/ui/button';
-import { Skeleton } from '../../shared/ui/skeleton';
-import EmptyState from '../../shared/components/EmptyState';
-import RevenueChart from '../charts/RevenueChart';
-import TopProductsChart from '../charts/TopProductsChart';
-import PaymentPieChart from '../charts/PaymentPieChart';
-import OrdersAreaChart from '../charts/OrdersAreaChart';
-import CashierPerformanceChart from '../charts/CashierPerformanceChart';
-import CategorySalesChart from '../charts/CategorySalesChart';
-import DistributorSalesChart from '../charts/DistributorSalesChart';
-import { formatCurrency } from '../../shared/lib/utils';
-import { useTranslation } from '../../shared/i18n/index';
+import { Card, CardContent, CardHeader, CardTitle } from '../../../shared/ui/card';
+import { Button } from '../../../shared/ui/button';
+import { Skeleton } from '../../../shared/ui/skeleton';
+import EmptyState from '../../../shared/components/EmptyState';
+import RevenueChart from './charts/RevenueChart';
+import TopProductsChart from './charts/TopProductsChart';
+import PaymentPieChart from './charts/PaymentPieChart';
+import OrdersAreaChart from './charts/OrdersAreaChart';
+import CashierPerformanceChart from './charts/CashierPerformanceChart';
+import CategorySalesChart from './charts/CategorySalesChart';
+import DistributorSalesChart from './charts/DistributorSalesChart';
+import { formatCurrency } from '../../../shared/lib/utils';
+import { useTranslation } from '../../../shared/i18n/index';
 import type {
   RevenueDataPoint,
   TopProduct,
@@ -29,7 +29,7 @@ import type {
   CashierPerformance,
   CategorySales,
   DistributorSales,
-} from '../../hooks/useDashboardData';
+} from '../hooks/useDashboardData';
 
 interface DashboardChartsProps {
   revenue: RevenueDataPoint[] | undefined;
