+++ client/src/features/analytics/components/charts/HeatmapChart.tsx
@@ -1,6 +1,6 @@
-import { useSettingsStore } from '../../shared/store/settingsStore';
-import { useTranslation } from '../../shared/i18n/index';
-import { formatCurrency } from '../../shared/lib/utils';
+import { useSettingsStore } from '../../../../shared/store/settingsStore';
+import { useTranslation } from '../../../../shared/i18n/index';
+import { formatCurrency } from '../../../../shared/lib/utils';
 import { useState } from 'react';
 
 interface HeatmapDataPoint {
