+++ client/src/features/analytics/components/charts/TopProductsChart.tsx
@@ -1,23 +1,21 @@
 import { BarChart, Bar, XAxis, YAxis, CartesianGrid, Tooltip, ResponsiveContainer } from 'recharts';
-import { useSettingsStore } from '../../shared/store/settingsStore';
-import { useTranslation } from '../../shared/i18n/index';
-import { formatCurrency } from '../../shared/lib/utils';
+import { useSettingsStore } from '../../../../shared/store/settingsStore';
+import { useTranslation } from '../../../../shared/i18n/index';
 import type { TooltipProps } from 'recharts';
 import type { ValueType, NameType } from 'recharts/types/component/DefaultTooltipContent';
 
-interface DistributorDataPoint {
-  distributor_name: string;
+interface TopProductDataPoint {
+  name: string;
   total_sold: number;
-  revenue: number;
 }
 
 interface CustomTooltipProps extends TooltipProps<ValueType, NameType> {
   isDark: boolean;
+  t: (key: string, params?: Record<string, string | number>) => string;
 }
 
-const CustomTooltip = ({ active, payload, isDark }: CustomTooltipProps) => {
+const CustomTooltip = ({ active, payload, isDark, t }: CustomTooltipProps) => {
   if (!active || !payload?.length) return null;
-  const data = payload[0].payload as DistributorDataPoint;
   return (
     <div
       style={{
@@ -27,25 +25,26 @@ const CustomTooltip = ({ active, payload, isDark }: CustomTooltipProps) => {
         borderWidth: 1,
         borderStyle: 'solid',
       }}
-      className="rounded-md p-3 shadow-lg space-y-1"
+      className="rounded-md p-3 shadow-lg"
     >
-      <p className="text-xs font-semibold font-data">{data.distributor_name}</p>
-      <p className="text-sm font-data text-gold">{formatCurrency(data.revenue)}</p>
-      <p className="text-xs font-data" style={{ color: isDark ? '#6B6B6B' : '#888888' }}>
-        {data.total_sold} items sold
+      <p className="text-xs font-data" style={{ color: isDark ? '#F5F0E8' : '#1E1E1E' }}>
+        {(payload[0].payload as TopProductDataPoint).name}
+      </p>
+      <p className="text-sm font-semibold text-gold font-data">
+        {payload[0].value} {t('charts.sold')}
       </p>
     </div>
   );
 };
 
-interface DistributorSalesChartProps {
-  data: DistributorDataPoint[];
+interface TopProductsChartProps {
+  data: TopProductDataPoint[];
 }
 
-export default function DistributorSalesChart({ data }: DistributorSalesChartProps) {
+export default function TopProductsChart({ data }: TopProductsChartProps) {
   const theme = useSettingsStore((s) => s.theme);
   const isDark = theme === 'dark';
-  const { isRtl } = useTranslation();
+  const { t, isRtl } = useTranslation();
 
   return (
     <div dir="ltr">
@@ -68,21 +67,20 @@ export default function DistributorSalesChart({ data }: DistributorSalesChartPro
             type="number"
             tick={{ fill: isDark ? '#6B6B6B' : '#888888', fontSize: 12 }}
             stroke={isDark ? '#1E1E1E' : '#E5E5E5'}
-            tickFormatter={(v) => formatCurrency(v)}
             reversed={isRtl}
           />
           <YAxis
-            dataKey="distributor_name"
+            dataKey="name"
             type="category"
             tick={{ fill: isDark ? '#F5F0E8' : '#333333', fontSize: 11 }}
             stroke={isDark ? '#1E1E1E' : '#E5E5E5'}
             width={isRtl ? 130 : 90}
             orientation={isRtl ? 'right' : 'left'}
           />
-          <Tooltip content={<CustomTooltip isDark={isDark} />} />
+          <Tooltip content={<CustomTooltip isDark={isDark} t={t} />} />
           <Bar
-            dataKey="revenue"
-            fill="#8B7355"
+            dataKey="total_sold"
+            fill="#C9A96E"
             radius={isRtl ? [4, 0, 0, 4] : [0, 4, 4, 0]}
             barSize={20}
             animationDuration={800}
